see #8465 - enable `-Xlint:cast` and fix associated warnings
git-svn-id: http://josm.openstreetmap.de/svn/trunk@7021 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/gui/layer/geoimage/CorrelateGpxWithImages.java
@@ -1174,7 +1174,7 @@ public class CorrelateGpxWithImages extends AbstractAction {
             WayPoint curWp, long curWpTime, long offset) {
         // Time between the track point and the previous one, 5 sec if first point, i.e. photos take
         // 5 sec before the first track point can be assumed to be take at the starting position
-        long interval = prevWpTime > 0 ? ((long)Math.abs(curWpTime - prevWpTime)) : 5*1000;
+        long interval = prevWpTime > 0 ? (Math.abs(curWpTime - prevWpTime)) : 5*1000;
         int ret = 0;
 
         // i is the index of the timewise last photo that has the same or earlier EXIF time
